Remove Location Check from gtrendsR
Remove the check that makes sure the location is real before passing the location to Google Trends. This was causing an issue because the DMA locations are no longer in the gtrendsR list of locations so it was failing een with real locations.
Now look for locations from the list of DMAs from the spring instead of pulling them from the countries dataset.